+++ packages/multicall/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @0xsodium/multicall
 
+## 1.7.0
+
+### Minor Changes
+
+- query token price use cache
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/abi@1.7.0
+  - @0xsodium/network@1.7.0
+  - @0xsodium/utils@1.7.0
+
 ## 1.6.0
 
 ### Minor Changes